[PB-1815] extract EditBillingDetailsModal to a component
