+++ src/app/newSettings/types.tsx
@@ -13,4 +13,14 @@ export interface Section {
   subsection?: string;
 }
 
+export type BillingDetails = {
+  address: string;
+  addressOptional?: string;
+  country: string;
+  city: string;
+  region: string;
+  postalCode: string;
+  phone: string;
+};
+
 export type Sections = Section[];
